Initial commit of ARPG unity (+ part of tilemap system)
The basic logic of these tilemaps is fairly simple:

User manipulates a grid of numbers
0 0 0 0
3 0 5 6
0 3 3 3

These numbers represent a combination of one or more tile + color, the first encoding step separates these two values and the various tile layers.

The second encoding step uses that information to find the correct sub-tile, and translates it into the final mesh data to be aggregated into one continuous map.